Move InputPassword.tsx and InputTeamName.tsx to root common folder
