+++ ui/src/Common/InputTeamName/InputTeamName.tsx
@@ -15,32 +15,38 @@
  * limitations under the License.
  */
 
-import React from 'react';
+import * as React from 'react';
+import Input from 'Common/Input/Input';
 
-import Input from '../../Input/Input';
-
-type Props = {
-	password: string;
-	onPasswordInputChange: (updatedTeamName: string) => void;
-	readOnly: boolean;
+interface Props {
+	teamName?: string;
+	required?: boolean;
+	onTeamNameInputChange: (updatedTeamName: string) => void;
 	invalid?: boolean;
-};
+	readOnly?: boolean;
+}
 
-function InputPassword(props: Props) {
-	const { password, onPasswordInputChange, invalid, readOnly } = props;
+function InputTeamName(props: Props) {
+	const {
+		teamName = '',
+		required,
+		onTeamNameInputChange,
+		invalid,
+		readOnly,
+	} = props;
 
 	return (
 		<Input
-			id="passwordInput"
-			label="Password"
-			type="password"
-			value={password}
-			onChange={(event) => onPasswordInputChange(event.target.value)}
-			validationMessage="8 or more characters with a mix of numbers and letters"
+			id="teamNameInput"
+			label="Team Name"
+			value={teamName}
+			required={required}
+			onChange={(event) => onTeamNameInputChange(event.target.value)}
+			validationMessage="Names must not contain special characters."
 			invalid={invalid}
 			readOnly={readOnly}
 		/>
 	);
 }
 
-export default InputPassword;
+export default InputTeamName;
